* WorldPacket now keeps opcode enum value. It is used for outputting opcode names to the logs * Change in SMSG_AURA_UPDATE. Thanks to LordJZ * Disabled SetPower/SetMaxPower for powers > 4 for now. * 6 opcodes

+++ src/server/game/Entities/Player/Player.cpp
@@ -22074,7 +22074,7 @@ void Player::SendAurasForTarget(Unit *target)
         uint32 flags = auraApp->GetFlags();
         if (aura->GetMaxDuration() > 0)
             flags |= AFLAG_DURATION;
-        data << uint8(flags);
+        data << uint16(flags);
         // level
         data << uint8(aura->GetCasterLevel());
         // charges
